+++ src/app.py
@@ -1,15 +1,22 @@
-import dash
+from dash import Dash, DiskcacheManager
 import dash_bootstrap_components as dbc
+import diskcache
 
-custom_css = '/assets/custom.css'
+custom_css = "/assets/custom.css"
 
 theme = dbc.themes.FLATLY  # https://bootswatch.com/flatly/
 
 
-app = dash.Dash(__name__,
-                external_stylesheets=[theme],
-                meta_tags=[{"name": "viewport", "content": "width=device-width"}],
-                suppress_callback_exceptions=True)
+cache = diskcache.Cache("./cache")
+background_callback_manager = DiskcacheManager(cache)
+
+app = Dash(
+    __name__,
+    external_stylesheets=[theme, dbc.icons.FONT_AWESOME],
+    meta_tags=[{"name": "viewport", "content": "width=device-width"}],
+    background_callback_manager=background_callback_manager,
+    suppress_callback_exceptions=True,
+)
 
 
 app.config.suppress_callback_exceptions = True