+++ src/sandboxing/sandbox_db2.py
@@ -38,30 +38,33 @@ query = user.tweets
 _vr_config = VrtoolConfig().from_json(Path(
     # r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\Case 7-2\config.json"
     # r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\TestCase1_38-1_no_housing_testingonly\vr_config.json"
-     r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\Case_31_1\config.json"
+    #  r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\Case_31_1\config.json"
     #  r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\Case_24_3\config.json"
     #  r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\Case_10_3\config.json"
+    r"C:\Users\hauth\bitbucket\VRtoolDashboard\tests\data\TestCase1_38-1_no_housing_no_dsn\vr_config.json"
+    # r"C:\Users\hauth\OneDrive - Stichting Deltares\Documents\tempo\VRM\38-1 test/config.json"
     # r"N:\Projects\11209000\11209353\B. Measurements and calculations\008 - Resultaten Proefvlucht\WDOD\10-3\database_vrtool_0_1_3_no_stix/config.json"
     #  r"N:\Projects\11209000\11209353\B. Measurements and calculations\008 - Resultaten Proefvlucht\WSS\31-1\database_vrtool_0_1_3_aangepast\config.json"
     # r"n:\Projects\11209000\11209353\B. Measurements and calculations\008 - Resultaten Proefvlucht\WSRL\24-3\02_Berekening\config.json"
     # r"N:\Projects\11209000\11209353\B. Measurements and calculations\008 - Resultaten Proefvlucht\WRIJ\47-1\database_vrtool_0_1_3\config.json"
     # r"N:\Projects\11209000\11209353\B. Measurements and calculations\008 - Resultaten Proefvlucht\WRIJ\47-1\database_vrtool_0_1_3\config.json"
 
-
 ))
 # _vr_config = VrtoolConfig().from_json(Path(__file__).parent.parent / "tests/data/TestCase1_38-1_no_housing/vr_config.json")
 t0 = time.time()
 _traject_db = get_dike_traject_from_config_ORM(vr_config=_vr_config, run_is_vr=1, run_id_dsn=2)
-data= _traject_db.serialize()
+data = _traject_db.serialize()
 export_to_json(data)
 t1 = time.time()
+
+# stop
 print(f"Time to get dike traject from ORM: {t1 - t0}")
 # _fig = plot_overview_map(_traject_db)
 
-# _fig = plot_dike_traject_reliability_initial_assessment_map(_traject_db, 2025, ResultType.RELIABILITY.name,
-#                                                             mechanism_type=Mechanism.REVETMENT.name, )
+# _fig = plot_dike_traject_reliability_initial_assessment_map(_traject_db, 2025, ResultType.INTERPRETATION_CLASS.name,
+#                                                             mechanism_type=Mechanism.SECTION.name, )
 _fig = plot_dike_traject_reliability_measures_assessment_map(_traject_db, 2025,
-                                                             result_type=ResultType.RELIABILITY.name,
+                                                             result_type=ResultType.INTERPRETATION_CLASS.name,
                                                              calc_type=CalcType.VEILIGHEIDSRENDEMENT.name,
                                                              colorbar_result_type=ColorBarResultType.MEASURE.name,
                                                              mechanism_type=Mechanism.SECTION.name,
@@ -82,3 +85,4 @@ print(f"Time to plot: {t2 - t1}")
 
 _fig.update_layout(mapbox=dict(accesstoken=get_mapbox_token()))
 _fig.show()
+